refactoring of local authentication
