+++ server/app/configure/authentication/local.js
@@ -2,85 +2,25 @@
 const passport = require('passport');
 const LocalStrategy = require('passport-local').Strategy;
 
+
 module.exports = function(app, db) {
 
     const User = db.model('User');
-    const Cart = db.model('Cart');
-
-    // When passport.authenticate('local') is used, this function will receive
-    // the email and password to run the actual authentication logic.
-    let strategyFn = function(email, password, done) {
-        User.findOne({
-                where: {
-                    email: email
-                }
-            })
-            .then(function(user) {
-                // user.correctPassword is a method from the User schema.
-                if (!user || !user.correctPassword(password)) {
-                    done(null, false);
-                } else {
-                    // Properly authenticated.
-                    done(null, user);
-                }
-            })
-            .catch(done);
-    };
+    const strategyFn = require('./fn/loc-strat')(db);
 
     passport.use(new LocalStrategy({ usernameField: 'email', passwordField: 'password' }, strategyFn));
 
     // A POST /login route is created to handle login.
     app.post('/login', function(req, res, next) {
 
+        let authCb = require('./fn/auth-cb')(req, res, next, db);
+
         if (req.user) {
             return res.sendStatus(400);
         } else {
-            let authCb = function(err, user) {
-
-                if (err) return next(err);
-
-                if (!user) {
-                    let error = new Error('Invalid login credentials.');
-                    error.status = 401;
-                    return next(error);
-                }
-
-                // req.logIn will establish our session.
-                req.logIn(user, function(loginErr) {
-
-                    let sendRes = function(obj) {
-                        return res.status(200).send({
-                            user: obj.sanitize()
-                        });
-                    }
-
-                    if (loginErr) {
-                        return next(loginErr);
-                        // We respond with a response object that has user with _id and email.
-                    } else {
-                        if (!req.session.hasOwnProperty('CartId')) return sendRes(user);
-                        Cart.findOne({
-                                where: {
-                                    id: req.session.CartId
-                                },
-                                include: [{ association: Cart.Product }]
-                            }).then(cart => {
-                                delete req.session.CartId;
-                                if (!cart) return;
-                                if (!cart.Items.length) {
-                                    return cart.destroy();
-                                } else {
-                                    return cart.update({ UserId: req.user.id });
-                                }
-                            })
-                            .then(() => sendRes(user))
-                            .catch(next)
-                    }
-                });
-
-            };
             passport.authenticate('local', authCb)(req, res, next);
         }
+
     });
 
     app.post('/logout', function(req, res, next) {
@@ -92,26 +32,7 @@ module.exports = function(app, db) {
     //A Post /create route for creating user accounts
     app.post('/createUser', function(req, res, next) {
 
-        let authCb = function(err, user) {
-
-            if (err) return next(err);
-
-            if (!user) {
-                let error = new Error('Invalid login credentials.');
-                error.status = 401;
-                return next(error);
-            }
-
-            // req.logIn will establish our session.
-            req.logIn(user, function(loginErr) {
-                if (loginErr) return next(loginErr);
-                // We respond with a response object that has user with _id and email.
-                res.status(200).send({
-                    user: user.sanitize()
-                });
-            });
-
-        };
+        let authCb = require('./fn/auth-cb')(req, res, next, db);
 
         User.findOrCreate({ where: { name: req.body.name, email: req.body.email, password: req.body.password } })
             .then(() => passport.authenticate('local', authCb)(req, res, next))
